requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.7
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
discord.py==2.2.2
distro==1.5.0
docker==5.0.0
docker-compose==1.29.1
dockerpty==0.4.1
docopt==0.6.2
flake8==3.9.2
idna==2.10
iniconfig==1.1.1
jsonschema==3.2.0
loguru==0.5.3
mccabe==0.6.1
multidict==5.1.0
packaging==20.9
paramiko==2.7.2
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
python-dotenv==0.17.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
texttable==1.6.3
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.5
websocket-client==0.58.0
yarl==1.6.3
youtube-dl==2021.4.26
